specifier.Grammar: move out of Parboiled
It's not related to Parboiled itself, just specifies the grammar.

commit-id:41d071c1
