+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/questions/BgpProcessPropertySpecifier.java
@@ -23,8 +23,8 @@ import org.batfish.datamodel.MultipathEquivalentAsPathMatchMode;
 import org.batfish.datamodel.answers.Schema;
 import org.batfish.datamodel.bgp.AddressFamily;
 import org.batfish.specifier.ConstantEnumSetSpecifier;
+import org.batfish.specifier.Grammar;
 import org.batfish.specifier.SpecifierFactories;
-import org.batfish.specifier.parboiled.Grammar;
 
 /**
  * Enables specification a set of BGP process properties.
