+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/questions/VxlanVniPropertySpecifier.java
@@ -14,8 +14,8 @@ import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.datamodel.answers.Schema;
 import org.batfish.specifier.ConstantEnumSetSpecifier;
+import org.batfish.specifier.Grammar;
 import org.batfish.specifier.SpecifierFactories;
-import org.batfish.specifier.parboiled.Grammar;
 
 /**
  * Enables specification of a set of VXLAN VNI properties.
