+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/parboiled/ParboiledAppSpecifier.java
@@ -16,6 +16,7 @@ import org.batfish.datamodel.applications.IcmpTypesApplication;
 import org.batfish.datamodel.applications.TcpApplication;
 import org.batfish.datamodel.applications.UdpApplication;
 import org.batfish.specifier.ApplicationSpecifier;
+import org.batfish.specifier.Grammar;
 import org.parboiled.errors.InvalidInputError;
 import org.parboiled.parserunners.ReportingParseRunner;
 import org.parboiled.support.ParsingResult;
