+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/parboiled/ParboiledEnumSetSpecifier.java
@@ -13,6 +13,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.specifier.EnumSetSpecifier;
+import org.batfish.specifier.Grammar;
 import org.parboiled.errors.InvalidInputError;
 import org.parboiled.parserunners.ReportingParseRunner;
 import org.parboiled.support.ParsingResult;
