+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/parboiled/ParboiledInputValidator.java
@@ -22,6 +22,7 @@ import org.batfish.referencelibrary.ReferenceBook;
 import org.batfish.referencelibrary.ReferenceLibrary;
 import org.batfish.role.NodeRoleDimension;
 import org.batfish.role.NodeRolesData;
+import org.batfish.specifier.Grammar;
 import org.batfish.specifier.Location;
 import org.batfish.specifier.LocationInfo;
 import org.batfish.specifier.SpecifierContext;
