+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/parboiled/ParboiledNodeSpecifier.java
@@ -8,6 +8,7 @@ import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.Set;
 import javax.annotation.ParametersAreNonnullByDefault;
+import org.batfish.specifier.Grammar;
 import org.batfish.specifier.NameNodeSpecifier;
 import org.batfish.specifier.NameRegexNodeSpecifier;
 import org.batfish.specifier.NodeSpecifier;
