+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/parboiled/ParboiledRoutingPolicySpecifier.java
@@ -8,6 +8,7 @@ import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.datamodel.routing_policy.RoutingPolicy;
+import org.batfish.specifier.Grammar;
 import org.batfish.specifier.NameRegexRoutingPolicySpecifier;
 import org.batfish.specifier.NameRoutingPolicySpecifier;
 import org.batfish.specifier.RoutingPolicySpecifier;
