+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/parboiled/ParserFilterTest.java
@@ -19,6 +19,7 @@ import static org.junit.Assert.assertThat;
 import com.google.common.collect.ImmutableSet;
 import java.util.Set;
 import org.batfish.common.CompletionMetadata;
+import org.batfish.specifier.Grammar;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
