+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/parboiled/ParserInterfaceTest.java
@@ -32,6 +32,7 @@ import org.batfish.referencelibrary.InterfaceGroup;
 import org.batfish.referencelibrary.ReferenceBook;
 import org.batfish.referencelibrary.ReferenceLibrary;
 import org.batfish.role.NodeRolesData;
+import org.batfish.specifier.Grammar;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
