+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/parboiled/ParserIpProtocolTest.java
@@ -12,6 +12,7 @@ import org.batfish.common.CompletionMetadata;
 import org.batfish.datamodel.IpProtocol;
 import org.batfish.referencelibrary.ReferenceLibrary;
 import org.batfish.role.NodeRolesData;
+import org.batfish.specifier.Grammar;
 import org.batfish.specifier.parboiled.Anchor.Type;
 import org.junit.Rule;
 import org.junit.Test;
