+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/parboiled/ParserLocationTest.java
@@ -23,6 +23,7 @@ import java.util.Set;
 import org.batfish.common.CompletionMetadata;
 import org.batfish.common.util.isp.IspModelingUtils;
 import org.batfish.datamodel.InterfaceType;
+import org.batfish.specifier.Grammar;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
