+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/parboiled/ParserNameSetTest.java
@@ -7,6 +7,7 @@ import com.google.common.collect.ImmutableSet;
 import org.batfish.common.CompletionMetadata;
 import org.batfish.referencelibrary.ReferenceLibrary;
 import org.batfish.role.NodeRolesData;
+import org.batfish.specifier.Grammar;
 import org.batfish.specifier.parboiled.Anchor.Type;
 import org.junit.Test;
 import org.parboiled.errors.InvalidInputError;
