+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/parboiled/ParserNodeTest.java
@@ -16,6 +16,7 @@ import java.util.Set;
 import org.batfish.common.CompletionMetadata;
 import org.batfish.common.autocomplete.NodeCompletionMetadata;
 import org.batfish.datamodel.DeviceType;
+import org.batfish.specifier.Grammar;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
