+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/parboiled/ParserUtilsTest.java
@@ -23,6 +23,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import java.util.Optional;
 import java.util.Set;
+import org.batfish.specifier.Grammar;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.parboiled.errors.InvalidInputError;
